+++ HISTORY.md
@@ -4,6 +4,7 @@
 
 ### 2025-04-03
 * tables: further rendering speed-up
+* objects, threads: new "attribute of color" reporter primitive in the pen category
 
 ### 2025-04-02
 * blocks, objects: added "color" selector to pen-attributes reporter